requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn

# Data sourcing
requests
python-dotenv
pyyaml
pyarrow
